add log in Rest ascured 29.03.2025
