+++ src/main/java/wbd/api/сlient/dto/AllOffersResponseDto.java
@@ -1,4 +1,4 @@
-package wbd.dto;
+package wbd.api.сlient.dto;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -10,7 +10,7 @@ import lombok.ToString;
 public class AllOffersResponseDto {
     private int id;
     private String title;
-    private CategoryResponseDto categoryResponseDto;
+    private CategoryResponseDto categoryDto;
     private double pricePerHour;
     private String description;
     private UserDetailsDto userFilterResponseDto;
