+++ src/main/java/wbd/api/сlient/dto/OfferDto.java
@@ -1,28 +1,17 @@
-package wbd.dto;
+package wbd.api.сlient.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
-import java.util.List;
-
 @Getter
 @Setter
 @ToString
-public class FilteredOffersResponseDto {
+public class OfferDto {
     private int id;
     private String title;
     private CategoryResponseDto categoryDto;
     private double pricePerHour;
     private String description;
-    private List<ImageDto> images;
     private UserFilterResponseDto userFilterResponseDto;
-
-    @Getter
-    @Setter
-    @ToString
-    public static class ImageDto {
-        private int id;
-        private String imageUrl;
-    }
 }
