+++ src/main/java/wbd/api/сlient/get/ApiClient_GetFilterOffers.java
@@ -1,15 +1,15 @@
-package wbd.api_client;
+package wbd.api.сlient.get;
 
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
 
-public class ApiClient_GetOfferById {
+public class ApiClient_GetFilterOffers {
 
-    public static Response getOfferById(int offerId) {
+    public static Response getFilteredOffers() {
         return RestAssured
                 .given()
                 .when()
-                .get("/offers/" + offerId)
+                .get("/offers/filter?cityName=all&category=all&keyPhrase=all")
                 .then()
                 .log().all()
                 .extract()
