+++ src/main/java/wbd/api/сlient/get/ApiClient_GetOfferById.java
@@ -1,15 +1,15 @@
-package wbd.api_client;
+package wbd.api.сlient.get;
 
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
 
-public class ApiClient_GetLocations {
+public class ApiClient_GetOfferById {
 
-    public static Response getLocations() {
+    public static Response getOfferById(int offerId) {
         return RestAssured
                 .given()
                 .when()
-                .get("/locations")
+                .get("/offers/" + offerId)
                 .then()
                 .log().all()
                 .extract()
