+++ src/main/java/wbd/web/core/ApplicationManager.java
@@ -1,4 +1,4 @@
-package wbd.core;
+package wbd.web.core;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.Point;
