+++ src/main/java/wbd/web/utils/RetryAnalyser.java
@@ -1,4 +1,4 @@
-package wbd.utils;
+package wbd.web.utils;
 
 import org.testng.IRetryAnalyzer;
 import org.testng.ITestResult;
