+++ src/main/java/wbd/web/web_pages/LoginPage.java
@@ -1,8 +1,8 @@
-package wbd.pages;
+package wbd.web.web_pages;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.WebDriverWait;
-import wbd.core.BasePage;
+import wbd.web.core.BasePage;
 
 public class LoginPage extends BasePage {
     public LoginPage(WebDriver driver, WebDriverWait wait) {
