+++ src/test/java/wbd/core/TestBaseUI.java
@@ -5,6 +5,7 @@ import org.slf4j.LoggerFactory;
 import org.testng.ITestResult;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
+import wbd.web.core.ApplicationManager;
 
 import java.lang.reflect.Method;
