+++ src/test/java/wbd/tests/rest_assured/GetLocationsTest.java
@@ -4,8 +4,8 @@ import io.restassured.response.Response;
 import org.testng.annotations.Test;
 import org.testng.asserts.SoftAssert;
 import wbd.core.TestBaseRA;
-import wbd.dto.LocationResponseDto;
-import wbd.api_client.ApiClient_GetLocations;
+import wbd.api.сlient.dto.LocationResponseDto;
+import wbd.api.сlient.get.ApiClient_GetLocations;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -45,7 +45,7 @@ public class GetLocationsTest extends TestBaseRA {
         // Проверяем, что каждое значение не пустое
         for (LocationResponseDto loc : locations) {
             softAssert.assertNotNull(loc.getCityName(), "City name should not be null");
-            softAssert.assertTrue(loc.getCityName().trim().length() > 0, "City name should not be empty");
+            softAssert.assertTrue(!loc.getCityName().trim().isEmpty(), "City name should not be empty");
         }
 
         softAssert.assertAll();
