+++ src/test/java/wbd/tests/rest_assured/GetOfferByIdTests.java
@@ -4,8 +4,8 @@ import io.restassured.response.Response;
 import org.testng.annotations.Test;
 import org.testng.asserts.SoftAssert;
 import wbd.core.TestBaseRA;
-import wbd.dto.FilteredOffersResponseDto;
-import wbd.api_client.ApiClient_GetOfferById;
+import wbd.api.сlient.dto.FilteredOffersResponseDto;
+import wbd.api.сlient.get.ApiClient_GetOfferById;
 
 import static io.restassured.RestAssured.given;
 
@@ -20,7 +20,7 @@ public class GetOfferByIdTests extends TestBaseRA {
         int offerId = 1;  // допустим, оффер с ID 1 существует
 
         Response response = ApiClient_GetOfferById.getOfferById(offerId); // получаем оффер по ID
-        logger.info("Response body: " + response.asString());
+        System.out.println("Response body: " + response.asString());
 
         SoftAssert softAssert = new SoftAssert();
 
@@ -28,9 +28,9 @@ public class GetOfferByIdTests extends TestBaseRA {
         softAssert.assertEquals(response.getStatusCode(), 200, "Expected status code 200");
 
         if (response.getStatusCode() == 200) {
-            logger.info("Тест прошел успешно: Код ответа 200 (OK)");
+            System.out.println("Тест прошел успешно: Код ответа 200 (OK)");
         } else {
-            logger.error("Ошибка: Получен ответ с кодом " + response.getStatusCode());
+            System.out.println("Ошибка: Получен ответ с кодом " + response.getStatusCode());
         }
 
         // парсим JSON в объект FilteredOffersResponseDto
@@ -77,18 +77,6 @@ public class GetOfferByIdTests extends TestBaseRA {
             softAssert.assertTrue(!offer.getImages().get(0).getImageUrl().isEmpty(), "ImageUrl не должен быть пустым");
         }
 
-        // запуск
-        softAssert.assertAll();
-    }
-
-    @Test
-    public void testGetOfferById_InvalidId() {
-        int invalidId = -1; // некорректный ID
-        Response response = ApiClient_GetOfferById.getOfferById(invalidId);
-
-        SoftAssert softAssert = new SoftAssert();
-        softAssert.assertNotEquals(response.getStatusCode(), 200, "Ожидается ошибка для некорректного ID");
-
         softAssert.assertAll();
     }
 
@@ -103,72 +91,25 @@ public class GetOfferByIdTests extends TestBaseRA {
         softAssert.assertEquals(response.getStatusCode(), 404, "Expected status code 404 for non-existent offer");
 
         if (response.getStatusCode() == 404) {
-            logger.error("Ошибка: Получен ответ с кодом 404 (Not Found)");
+            System.out.println("Ошибка: Получен ответ с кодом 404 (Not Found)");
         }
 
         softAssert.assertAll();
     }
 
-    @Test
+    @Test  // написать баг-репорт
     public void testGetOffersByInvalidId_400() {
         // запрос с несуществующим или неправильным параметром ID
         Response response = given()
                 .when()
-                .get("/offers/invalid-id")
+                .get("/api/offers/invalid-id")
                 .then()
                 .statusCode(400)  // проверка на статус 400
                 .log().all()
                 .extract().response();
 
         if (response.getStatusCode() == 400) {
-            logger.info("Тест прошел успешно: Получен ответ с кодом 400 (Bad Request)");
+            System.out.println("Тест прошел успешно: Получен ответ с кодом 400 (Bad Request)");
         }
     }
-
-    // тест для неавторизованного пользователя
-    @Test
-    public void testGetOfferWithoutToken() {
-
-        Response response = given()
-                .when()
-                .get("offers/1")  // запрос без авторизации
-                .then()
-                .statusCode(200)  // ожидаем успешный статус 200
-                .log().all()
-                .extract().response();
-
-        // проверяем, что ответ не содержит данных пользователя (имя, фото, местоположение)
-        String responseBody = response.getBody().asString();
-        SoftAssert softAssert = new SoftAssert();
-
-        // Ожидаем, что в ответе будет информации о пользователе
-        softAssert.assertTrue(responseBody.contains("firstName"), "Имя исполнителя должно быть видно неавторизованному пользователю");
-        softAssert.assertTrue(responseBody.contains("lastName"), "Фамилия исполнителя  должна быть видна неавторизованному пользователю");
-        softAssert.assertTrue(responseBody.contains("profilePicture"), "Фото исполнителя  должно быть видно неавторизованному пользователю");
-        softAssert.assertTrue(responseBody.contains("locationDto"), "Местоположение исполнителя должно быть видно неавторизованному пользователю");
-        softAssert.assertTrue(responseBody.contains("title"), "Название предложения должно быть видно");
-        softAssert.assertTrue(responseBody.contains("categoryDto"), "Категория предложения должна быть видна");
-        softAssert.assertTrue(responseBody.contains("pricePerHour"), "Цена за час должна быть видна");
-        softAssert.assertTrue(responseBody.contains("description"), "Описание предложения должно быть видно");
-
-        softAssert.assertAll();
-    }
-
-    @Test  // нужен багрепорт
-    public void testServerError() {
-        // симулируем ошибку на сервере, например, при вызове неправильного endpoint или когда сервер не может обработать запрос
-        Response response = given()
-                .when()
-                .get("/offers/trigger-server-error")
-                .then()
-                .statusCode(500)  // проверка на статус 500
-                .log().all()
-                .extract().response();
-
-        if (response.getStatusCode() == 500) {
-            logger.info("Ошибка: Получен ответ с кодом 500 (Internal Server Error)");
-        }
-    }
-
-
 }
